refactor: rename createCustomer to signup for better consistency
