+++ src/components/blocks/UserProfile/parts/AddressList/address.validation.ts
@@ -1,5 +1,5 @@
 import * as yup from 'yup';
-import type { CountryCode } from '../../../../../api/createCustomer/createCustomer.types';
+import type { CountryCode } from '../../../../../api/signup/createCustomer.types';
 import type { AddressType } from '../../../../../api/profile/profile.types';
 
 const ruleForStreet = yup
